Add data migration to fix references to LGSL 1473
As noted in previous commit, LGSL 1473 was added to local_services.csv in
error; it should have been 1743.  This was corrected in
5411e984203fe57149089f06961ca76fe201db46, but not before an edition was
created that referenced this.  That edition is now archived but it's still
better if it's a valid edition so we change it to reference the correct
LGSL code: 1743.