Provide rake task to remove old services
Now that we've trimmed the local_services.csv to only include the services
that we actually use, we want to clean up the instances of LocalService in
the db.  The cleaner will remove any local service whose LGSL code is not
in the CSV and also that is not used by any local transaction editions (to
avoid us having models in the db that are invalid).

Having run this it means that it is no longer possible to create a new
local transaction for a new service, as the lgsl_code in the local
transaction must be one that is stored in a LocalService instance. To add
a new one will require us to add a new row to the CSV, deploy it and run
the import.  (Which might have been the case anyway if the desired LGSL
was not one of the ~250 in the previous version of the CSV).