Migrate organ donor registration promotion data
We no longer want to provide the api for accessing organ donor
registration promotions directly and we also don't want the presentation
toggles storage to store this promote differently.  This migration
removes all the organ donor registration promotion data and re-saves it
in the promotion choice hash instead.  We also make sure that all the
done pages have the new promotion choice hash in presentation toggles so
that everything is consistent..