Add LGSL 107 as duplicate of 230
We want to migrate from using LGSL 230 for the sheltered housing transaction
to using LGSL 107 instead. As a first step we'll add 107 as an an "enabled"
service before moving everything else over.  Once 107 is fully live we can
retire the LGSL 230 service.

This work is tracked in the following Trello story:
https://trello.com/c/Zb42F6Jq/417-switch-local-transaction-for-apply-for-sheltered-housing-from-one-lgsl-to-another-3