Make 20160805145635_remove_local_interactions.rb runnable
The commits after this migration removed the LocalAuthority model
entirely so this migration can no longer run successfully.  The next
migration also removes the local_authorities collection so the work
this migration does is superceded by that migration.  Commenting out
the broken reference to LocalAuthority seems like the best approach
to making it runnable.