Fix bug when saving external related links
These changes are necessary for avoiding the app to crash when someone tries to
save without adding new links to an item without links.

The bug is caused by the fact that the artefact item is not generated when no
links are added.

This is a dirtish fix, it would be nice to merge the tagging tabs together and
hide/deactivate the save button when no changes are applied to the edition.