Include format filtering in the user search UI
We also take the opportunity to normalise the UI on this view by giving
it the same labels on the filters as the root page.  Apart from the user
filter which on this page does not just search for documents assigned
to the user, but also documents they have been involved with (requested
changes, etc...).

This lets us use the same filter_by_xxx test helpers in the root_overview
and user_search integration tests.  With one change: we remove the
`click_on "Drafts"` action from the filter_by_user helper.  None of the
existing tests for root_overview failed, and it seems to be a surprising
step (I've asked to filter by user, so why am I also filtering by drafts
immediately afterwards?).