Really use GdsApi.config.always_raise_for_not_found
The gds_api.rb initializer is overwritten on deploy by
[files from alphagov-deployment](https://github.gds/gds/alphagov-deployment/blob/master/publisher/initializers_by_organisation/integration/gds_api.rb)

These files in alphagov-deployment don't have the `always_raise_for_not_found`
configuration set.  This means that it's only used in dev.  I recently tried
to create a new `Artefact` in Panopticon in dev (which worked), but I was
subsequently unable to edit the created `Edition` in Publisher as Publishing-api
returned a 404 for `Services.publishing_api.get_links` on it.

Production environments are currently unaffected, but if we fix the
alphagov-deployment files to also have this configuration before this fix is
merged we'll be unable to edit `Editions` on newly created `Artefacts`.

The production initializers also try to use environment variables, but this is
not backed up by Puppet or the deployment repo, so this is something that we'll
fix in the near future.

[Trello story](https://trello.com/c/v7pdphVZ/83-publisher-api-adapters-upgrade-is-incomplete)