Skip review button should be available permitted authors of content
The original story for this feature mentioned placing the button within
the review buttons visible when viewing another editor's work when in
review. The skip review feature should mean that an author with the correct
permissions can skip the review stage of their own work which means the
relevant button needs to appear when the user is viewing their own work.