Move publishing api calls to controllers
This commit aims to solve the following problems:
- Publisher creating a draft content item in PublishingAPI after making a 'publish' request to PublishingAPI
- Publisher sending the wrong content to PublishingAPI before a 'publish' request
- Publisher sending multiple redundant 'put_content' requests as part of its Edition-related workflow

Before this commit, the Edition model had an after_save hook that would make a 'put_content' request to PublishingAPI each time an Edition was saved. This would happen multiple times per state change for an edition (see workflow.rb) due to the state machine and extra calls to record users (see denormalize users). Since in the PublishingAPI we do not yet model the Publisher workflow, doing a put_content for these state changes is redundant, however it is possible to change edition content while also changing state (e.g when requesting 2i without clicking save directly ) so we cannot remove all the calls.

Also when an edition is published, each previous edition for that artefact would be set to 'archived', thereby triggering the save method and put_content calls to PublishingAPI. This led to the wrong content being in the PublishingAPI when 'publish' was finally called. Also, after the publish, the published edition in Publisher would be saved again, causing another put_content which would create a new draft content_item in the PublishingAPI.

To fix the issues described above, we remove the after_save hook from the Edition Model and call the put_content commands from the relevant controller methods instead. This makes it more obvious at which point Publisher communicates with the PublishingAPI and resolves the issues described above.