Remove call to Panopticon on Edition.save
Publisher has been pushing an update to the Panopticon JSON API every
time an edition was saved. This performed two actions: it updated the
artefact from the edition, and it pushed change notifications out to
the router-api, rummager and the publishing-api.

We have changed the behaviour so that publisher no-longer calls
Panopticon, and instead performs the same actions internally.

We have introduced a new method on the local monkeypatch of the Artefact
model - update_from_edition - which makes the changes that would have been
persisted by the API call to Panopticon.

There's also a new method on the local Edition monkeypatch.
'notify_publishing_platform_services' which makes the appropriate API calls
to register changed content with the publishing platform.

There are various Rake tasks for publishing and modifying Publisher content
in some way, and these too have been updated to use the new methods.