Fix Rummager search bug
We discovered this bug in Errbit
(https://errbit.publishing.service.gov.uk/apps/5304e0520da11511b0000011/problems/58872a52657863786d413b00)
where when a piece of content is unpublished, the item is not removed from Rummager.

This was to do with a deprecated `delete_content!` method. We also
removed the functionality that retried deletion.