Create presenter for new Help schema
As part of mainstream format migration, we've created a content-schema
for help pages.

To support this, and future content schemas, we've created a small
number of classes:

A Format::HelpPagePresenter which converts a HelpPageEdition into the
correct format for Publishing-API
An EditionPresenterFactory which chooses and configures the correct
Format Presenter class for the type of content.