Add segments column to multipart redirects export task
[This PR](https://github.com/alphagov/publisher/pull/552) changed the
export of the multipart redirects to remove linking to a page fragment
identifier. This was felt to not be robust as content changes regularly
and these identifiers may not exist in the future.

An alternative to this is to add the `Segments` column to the exported
csv. Router-API would then [take care](https://github.com/alphagov/router-api/blob/8828caf69a69533023a1f6fe68bc7276ccc42b6f/app/models/route.rb#L77) of whether or not to preserve or
ignore the segment redirect. Broken page fragment identifiers would
simply send the user to the top of the page.