Remove publish to publishing api from edition callback
When an edition's state changed to 'published' as part of the state machine transitions, Publisher would make calls to PublishingAPI and Rummager. We had previously separated any `put_content` calls to PublishingAPI from the edition and put them into the controllers instead because it's easier to follow when these calls happen. This does the same for the publish-related calls. We also create some service objects to group related calls together.