Add fact_check_ids to draft content sent to publishing-api
So that un-authenticated users can fact-check draft content, we must
ensure that documents in the draft-content-store have the required fields.

The fact_check_ids we set are only present for content that is expecting
a fact-check, aka in the fact_check, fact_check_received, or ready
workflow state.

We have created a deterministic hashing process that means a URL can be
shared that provides access to a specific draft edition. Future editions
will not be accessible via this URL, and the content designer will need
to request further fact-checks via the Publisher UI. This is no different
to the current editorial experience, except that fact-check URLs will no
longer work once an edition has been published.

The #fact_check_id method enables us to tie access control of draft
content to the publisher Edition ID, and encapsulates this logic  without
forcing the developer to manage the lifecycle of a persisted attribute.
There is already enough platform complexity sprinkled throughout the app.

The Edition#fact_check_id method creates a just-unique-enough "uuid"
which is absolutely not guaranteed to be universally unique, but only
needs to be unique across the range of Editions at a given URL. It was
inspired by/shamelessly ripped off from the SecureRandom module in stdlib.
http://ruby-doc.org/stdlib-1.9.3/libdoc/securerandom/rdoc/SecureRandom.html#uuid-method