Enable non-authenticated users to view draft content for fact-checks
Using JSON Web Tokens we can allow un-authenticated users to view
specific pages in the draft stack.

This change adds the JWT token to the generated URL in the fact-check
invitation emails. It will allow anyone with that URL to access the
specific draft edition it refers to. The URL will cease to work once
the edition has been published, and won't work for future editions.