Disable govuk linter for downstream schema tests
There were CI failures against upstream govuk-content-schema branches
because of the configuration of the linter in this Jenkinsfile which was
running against the diff of deployed-to-production and origin/master.

There's really no need for the linter to run when we're testing schema
compatibility, so we skip it in this case.