Create new migrated? test that doesn't rely on Artefact#kind
We discovered an error when republishing Answer content. Namely,
our assumption that Artefact#kind will always equal the Edition type
was wrong. When an Edition changes format, the Artefact#kind doesn't
change with the edition.

There were several places where we enshrined this false equivalence
in the application, and this is an attempt to replace them with logic
predicated on the type of Edition class.

Normally I wouldn't advocate switching on class name in a super-class
method, however there are mitigating factors.  For the time being, the
Edition and all subclasses are defined in the govuk-content-models gem
and there's significant friction republishing that project everytime we
migrate a new format. It's expected that once the mainstream formats have
all been migrated to rendered, and content-api has been decomissioned,
then these shared models can be repatriated to the Publisher repo for ever
more.