Send document_type based on edition type
Previously this was determined by the 'kind' attribute on the edition's artefact. However when an edition was converted into another type (eg from Answer to Guide), the artefact's 'kind' would not update. When sending this information to the publishing api, this could end up being misleading. Instead we chose to send the document type to be the same as schema_name (since we don't use sub document types so far).