Add Completed Transaction presenter
We are using "promotion" in the presenter, veering away from the model
attribute name "promotion_choice" as the promotion choice is being made in
publisher, but will no longer change after being presented to the publishing
platform. For similar reasons we are using "category" instead of "choice".

We also no longer sending 'none' promotions to the Publishing API.