Add fact check token to preview link for preludial states
Users of Publisher want to provide access to the draft content, even
when its not in the fact check state. Therefore, add the fact check
token to the Preview button link.

As the preview link now includes the token, remove the
fact_check_edition_path function and changes uses of it to use
preview_edition_path.

Alter the model to allow for the generation of the fact check id
regardless of the state.

Note that the Publishing API currently strips the access limited
information when sending published content to the draft content
store. This means that the token won't have any effect, therefore to
avoid confusion, don't include the token if this is the case.

Change the edition in the GenericEditionPresenterTest to be an
unmigrated document_type, as then the fact_check_id does not appear in
the Publishing API payload.