Use the artefact "kind" as document type
For non-migrated editions we're currently sending the schema name
(`generic_with_external_related_links`) to the content store as
document type. This is affecting the analytics gathering, which uses
document type.

This separates the two concepts by having a separate method for the
type. This will make sure we won't confuse the two anymore.

As noted in https://github.com/alphagov/publisher/pull/566,
`@artefact.kind` isn't always the correct type (the kind doesn't change
if the format changes). This will be solved by migrating all the
documents. In the meantime, the `kind` is preferable to
`generic_with_external_related_links`.