Adapting Downtime to the new publishing platform
The Downtime feature could not be migrated without substantial changes,
as it was tied to artefact-level functionality baked into the content-api
which cannot be reproduced with the existing publishing-api workflow assumptions.

Here we have implemented a system which essentially bypasses the publishing-api
workflow to update the published version and latest draft, without publishing any
existing draft edition. It relies on Sidekiq to schedule the publication and
expiration of Downtime messages.

Along the way we fixed an almost-bug in the implementation of the Rails `time_select`
and `date_select` form helpers, which was impacting on our test data. Without the
`ignore_date` flag being set, `time_select` will default to adding three hidden fields
for the date component. When used alongside the `date_select` date picker, this creates
duplicate form-data, which became very confusing when we came to test the interaction
between the rendered page and the controller.