Bump govuk_content_models to 43.2.0
We bump govuk_content_models as Programme is now a retired format, and this
will ensure that users are no longer able to create a Programme artefact or
edition, and also are no longer able to convert another format to a Programme.
Additionally, users will be able to see that Programme is retired in the Format
dropdown menu, as well as when they view a Programme edition.

Trello: https://trello.com/c/aonp9YbK/675-retire-programme-format-2