Bump govuk_content_models to 44.1.0
We bump `govuk_content_models` as Campaign is now a retired format, and this
version will ensure that users are no longer able to create a Campaign artefact
or edition, and also are no longer able to convert another format to a
Campaign.  Additionally, users will be able to see that Campaign is retired in
the Format dropdown menu, as well as when they view a Programme edition.