Fix deprecation warnings
Upgrading to Rails 5 requires upgrading the Airbrake gem to avoid deprecation
warnings about middleware. However, Airbrake 5 breaks some compatibility with
our Errbit installation, namely that deployments are no longer recorded.
Instead of upgrading to a released version of the Airbrake 5 gem, we use a
branch in GDS's fork of Airbrake, which is version 4.3.8 of the gem with a
small change to silence the warnings.
- https://docs.publishing.service.gov.uk/manual/upgrade-rails.html#airbrake

In Rails 5, whenever we define a belongs_to association, it is required to have
the associated record present by default. This avoids this behaviour.

Using positional arguments in functional tests has been deprecated, in favor of
keyword arguments, and will be removed in Rails 5.1.  We therefore need to wrap
params in a params hash for controller calls

Remove deprecated before_filter and after_filter and replace with before_action
and after_action

Replacing deprecated "render :text"

Removing deprecated silence_stream method

"DEPRECATION WARNING: `:nothing` option is deprecated and will be
removed in Rails 5.1. Use `head` method to respond with empty response body."