Upgrade Kaminari
Bumping Kaminari to 0.17 fixes errors relating to sanitising Rails params.

https://github.com/kaminari/kaminari/pull/766

Adding the kaminari-mongoid gem fixes the following deprecation warning:

```
DEPRECATION WARNING: Kaminari Mongoid
support has been extracted to a separate gem, and will be removed in the next
1.0 release. Please bundle kaminari-mongoid gem.
```

Initializing Kaminari resolves issues with `undefined method `paginate'` errors,
that seem to be related to the models residing in a gem as opposed to natively
within the app.