Permit params explicitly
Nested attributes need to be permitted as an array, so this changes permissions
for topics, organisations and mainstream_browse_pages for
tagging_update_form_params in editions_controller. It also turns params to
hashes when relevant, and removes redirect_to:back, which was deprecated.

We also explicitly allow params for Downtimes Controller