Fix linter issues in the entire project
There are inconsistencies between running the linter locally or on
Jenkins. Sometimes on Jenkins it reports problems on files not edited,
which is not great.

In order to get around that problem, I've fixed all linter errors on
this project in 2 ways:

- first I can the linter with `-a` option and let it fix most of the
  issues automatically;
- finally, I manually fixed the remaining 50 issues.

This will hopefully help reducing the noise going forward.