Remove testing of the performance of the homepage
Upgrading to Rails 5 surfaces an inconsistency between rails-perftest and
rubyprof. Rails-perftest provides arguments to a rubyprof method that expects 0
or 1 arguments, which is made up of the params + some options that have been
merged into them as a hash. However given in Rails 5 params are no longer a
Hash, the merging isn't working as expected, and that means rails-perftest is
sending 2 arguments: the params and the options hash separately. As a result
rubyprof is failing and the performance test is breaking.

Given these gems are only used by one performance test of the homepage, which
is publisher-facing only, we reckon removing it is fine for now.