Bumping test gems
Bumping test gems as far as possible to hopefully make the 5.1 upgrade easier.

I'm removing a few gems that don't appear to be used.  SOme of them (such as
capybara-screenshot and launchy) look like they might be handy for development,
but in the interests of keeping the Gemfile maintainable, I'm removing them for
now.

I'm not sure why the RSpec CI reporter was added as we don't have any RSpec
tests in Publisher.  I suspect it was done in error at the time of everything
moving to the new Jenkins.