Incorporate `govuk_content_models`
The `govuk_content_models` gem used to share models between several
apps. Gradually the number of applications using the gem reduced, until
`publisher` was the final application using it.

Given that no other applications are using the models, there’s no need
to share the models, and hence the can be retired if we move its
functionality into Publisher.

This change is a copy-and-paste incorporation from
`govuk_content_models`.