Add test to keep promotion values in sync with schema
We want to make sure that the completed transaction edition and presenter
don't allow values for promotion that are not accepted by the schema as
this would result in rejection from the publishing-api.  We use the
govuk_schemas gem to pull in the completed_transaction schema and compare
`FormatsCompletedTransactionPresenter::PROMOTIONS` and
`CompletedTransactionEdtion::PROMOTIONS` to the definition of promotions
in the schema.

We can't do a `x must equal y` test here as this would break deploys of
both publisher and govuk-content-schemas if we ever want to change the
allowed values.  When we make a change to schemas to add a value we run
the tests of publisher against this change, using the production deploy
version of publisher.  This test run would fail because the two
definitions wouldn't match.  If instead we make a change to publisher
first it will run its tests against the production version of the schema.
This test run will also fail because the two definitions wouldn't match.

The solution is to make the test check that there's no value in the
publisher definition that is not also in the schema definition.  This
means publisher will always be a little bit out of date when we add a new
value and a little bit into the future when we remove a value.  However we
will be able to deploy publisher and schemas without failed tests.