Enable editing of Transaction Start button text
This change allows publishers to choose the text that appears in the
green “start” button on a Transaction. This behaviour is basically a
direct read-across from the similar functionality in Simple Smart
Answers.

Publishers will be able to select from a drop-down, which only contains
two choices:

- “Start now”
- “Sign in”

“Start now” is the text that is currently displayed, and remains the
default.

“Sign in” is the new copy requested specifically for the Universal
Credit Sign In page.

We use a drop-down to restrict the choices. Copy will need to go through
approval to be included in the drop-down, and having a list of these
items will help to keep consistency across GOV.UK (for example, we will
not have variations like “Sign in”, “Signin”, “Log in”, etc.).