Remove Need API
Need API is being retired so we're removing this rake task as it won't work anymore.

The rake task was initially used to export needs from Need API to the publishing API. This has now been done so that rake task is not needed anymore.

For: https://trello.com/c/ClCFMBlG/217-retire-need-api