Make sure editions have artefacts with a matching kind in tests
We don't have anything that makes sure an artefact's kind matches up with
the format of its edition, but it's useful to make sure that they do in
our tests, so that if we do add this later we don't break a bunch of tests
unexpectedly.

The one gotcha here is that the slug validator for a done page artefact
makes sure that it starts with `done/` so our factory also needs to make
sure that is true.  Interestingly, the validations on the edition do not
enforce this which feels strange, but I guess editions are cloned from
their artefact and the artefact is created first.