When changing an edition's format, also change its artefact's kind
This means that we do a better job of making sure that artefact kinds and
edition formats stay in sync as they get changed around.  It's unlikely
the discrepancy actually causes any problems, but it's surprising and may
bite us in the future if documents become migrated to another app and the
editions are purged, but the artefacts remain.