Add exact_route? to Artefact
If there is a latest_edition we delegate to the implementation on that,
but there are many Artefacts without editions so we can't always do this.
Mostly these represent artefacts created as placeholders for content not
edited by publisher (smart-answers, manuals, custom apps like calendars
etc...), but there are also a handful of artefacts that are "publisher"
formats, but do not have any editions (content created in panopticon, but
no draft ever saved in publisher most likely).  For the latter we can
rely on the `Edition::EXACT_ROUTE_EDITION_CLASSES` to tell us if the
artefact is an exact route or not, for the former we can check if the
artefact has any `prefixes` - if it does it's not an exact route, if it
doesn't it is.

Why might we want to do this?  Well, we're shortly going to remove all
the business support editions and want to use the unpublish service to
redirect them via the publishing-api properly (currently this is done
via router-data).  This service currently works with an Artefact, but
needs to know if it's an exact route which only works with if the
artefact has a latest_edition.  Once those editions are gone we won't
be able to run that anymore, but if artefacts know if they are exact
routes or not themselves then we don't have that problem.