Change UnpublishService to ask the artefact about exact routes
In the previous commit we added an `exact_routes?` method to `Artefact`
so that it does not need to rely on its latest edition (which it may not
have) for this info.  Now that `UnpublishService` uses this method it no
longer needs an artefact to have an edition, which means it can be used
on all the artefacts in the db.