Remove monkey patch for mongoid dirty tracking of arrays
The code in here no longer does what we expect because Mongoid doesn't
have a Mongoid::Dirty module anymore.  It was changed to
[Mongoid::Changeable in v4][move] (we currently use v6) and the specific
issue we were tracking was [fixed in v3.1.0][fix].  The tests are no
longer useful because mongoid itself is fixed, but also because when we
remove business support there will be no edition subclass or class that
includes the workflow module that has an Array field so there's no object
we can instantiate in the test to check the specific thing we need.

[move]: https://github.com/mongodb/mongoid/commit/c16ff80c2f52d0de2d53df4bad3f313db4e5550e#diff-6a93775129aea909d2a0d5b3a37c6f80
[fix]: https://github.com/mongodb/mongoid/commit/b78e7a9286330b3223d000073ee9e8e1eb5790bc