Start migration to deal with business support artefacts
We'll deal with live, archived, and draft types separately as we want to
do different things to them all.  The next few commits will introduce the
code for handling each state.

If there are artefacts in states other than live, archived or draft, we halt
the migration as the data is not as we expected.