Deal with draft business support artefacts
In our investigations with the data, none of these has any editions, an
entry in router for their slug, or an entry in publishing-api for their
content_id.  If this is truly the case then we can just delete them and
pretend they never existed.  We guard against the possibility that this
isn't the case with live data and halt the migration before destroying
all the artefacts.  Most likely these were created in panopticon but
never saved in publisher so have been hanging around in a limbo state
for ages.