Deal with live business support artefacts
Our goal is to use the `UnpublishService` to archive all these artefacts
and send an unpublish notice with a redirect to the publishing-api.  All
but 3 of these artefacts have already been redirected via router-data and
so we can get the redirect url from there.  For the 3 that haven't one we
need to treat specially, 'jobcentre-plus-vacancy-filling-system-uk', the
other two we can redirect to the new business finance support finder route
at '/business-finance-support'.  Dealing with
'jobcentre-plus-vacancy-filling-system-uk' is more complex because it has
already been turned into a redirect with a new content_id via
short-url-manager.  Instead of issuing a redirect we check that the state
of the world is as we expect (there's a redirect in router, publishing-api
doesn't think our content_id owns the path, the object that does is a
redirect) and halt the migration if it's not.  We also halt the migration
if the state of the world is not as we expect for any of the other live
artefacts (there's no route in router, publishing-api thinks we own the
path, and the object in publishing-api is a business_support document).