Fix equality typo in republishing business_support migration
We had `=` instead of `==` when detecting the
`jobcentre-plus-vacancy-filling-system-uk` artefact which means we picked
the wrong artefact to check the redirect against.  Unluckily for us this
artefact's routes had the same conditions as the
`jobcentre-plus-vacancy-filling-system-uk` one so picking the wrong
artefact didn't cause an exception to be raised.  Ultimately what this
does is mean we filtered this artefact out of the list of live ones to
process and don't redirect it properly.  Replacing the `=` with `==`
means we pick the correct artefact for
`jobcentre-plus-vacancy-filling-system-uk` and only filter it out.