Update the button text when initialising a collapsible group
The text of the toggle is controlled when we change the state of any of
the collapsible things, but we weren't checking it at intialisation time
so it could be out of date.  For example if we render a list of 5
collapsibles but add the 'in' class to the first one so we start with it
open, but the rest closed then we want the toggle text to have the
collapse text, as that's the behaviour of the JS in that case.  Similarly
if we render so all are closed then we want the toggle text to be the
expand text.  Should we decide to change the behaviour of the collasible
group from always collapsing if any are open, to always expanding if any
are closed, this will protect us from the get go without having to
remember to change the starting text in our HTML.