Add collapsible-group toggle to the history and notes tab
It's already using the bootstrap collapse behaviour so it's easy for us to
add the collapsible-group module and toggle link as well to get quick
expand all/collapse all behaviour for this tab.  One downside is that
because we start with the first edition notes open, the default behaviour
of the toggle is to collapse all of the notes, rather than what'd we'd
probably want, which is to open the rest of them.