Add links method to service_sign_in_presenter
We need to add a parent link to the `service_sign_in` documents so we can refer
back to its parent transaction or guide page. Parent links are link-set links
which are sent to Publishing API using `PATCH link-set` and so it needs to be
separate from the rest of the payload. Therefore we add a new method called
`links` which will enable us to send parent links to the Publishing API.