Introduce "Resend fact check" workflow action
This isn't like other workflow actions because it doesn't move you through
the state machine.  You can do it when you are in "fact-check" mode, but
all it does is resend the previous fact check email.  This means we need
to define our own `can_resend_fact_check?` and `resend_fact_check` methods
on `Workflow`.  `can_resend_fact_check?` returns true if we are in the
fact check state, and `resend_fact_check` is a no-op method that always
returns true.  This is because the only thing we want to do is send an
email and this is all handled by the
`GovukContentModels::ActionProcessors` and `NoisyWorkflow` mailer. It's a
little confusing, but it'd be worse if this "action" method was different
to the others, and was the only place where email was sent without going
through the processors.

The `NoisyWorkflow.resend_fact_check` method is a little unusual in that
it's not a standard mailer action.  To resend a fact check we need to find
the most recent `send_fact_check` action and re-use the details in it to
send the fact check, if we can't find it we can't send an email.  This is
wrapped up in `NoisyWorkflow.resend_fact_check` - if the supplied action
object isn't a `resend_fact_check` or we the previous status action for
the edition isn't a `send_fact_check` then we return a `NoMail` instance.
A `NoMail` is a null object that provides enough of the actionmailer
delivery API that our calling code doesn't have to understand what to do
with a mail that can't be sent.  If everything is as we expect then we
call the `request_fact_check` method on the previous status action and
let actionmailer do it's thing to return a real mail object that can be
delivered.