Add Service Sign In Yaml Validator
We have a rake task that takes a provided YAML file containing
`service_sign_in` content and publishes it to the Publishing API.  We add a
Service Sign In Yaml Validator to check that the file is valid before
attempting to publish the content.  As a first step, we load the YAML file and
return it as a hash.