Check for missing choose_sign_in option fields
After checking if the required `choose_sign_in` fields are present, we need to
check if the required fields for `choose_sign_in` `options` are present.
Before this, we check that `options` has an array and log an error if not.
Note that in the YAML file users can provide either an internal `slug` or
external `url`, which is why we check for the presence of either.