Pass "locale" to service_sign_in publish call
Publishing-api assumes that the locale is "en" if you do not
explicitly pass the locale in the publish call. As we are also
publishing Welsh content, we should always pass the locale of the
content item.

We have to pass an update_type too because of the order that
publishing-api expects to see the params, however as the actual value
of update_type used by publishing-api comes from the `put_content` we can
set this value to `nil`.