Add edition progress errors to flash during update
Progressing an edition doesn't always succeed and if we do that via an
update action the error message is lost so the user may not understand
that their action has failed.  We follow the lead of the `progress`
controller action and check the return value of `progress_edition` and
set the flash accordingly.